requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-docx==1.1.2
SpeechRecognition==3.10.0
requests==2.32.3
pydub==0.25.1
